Cache Keras callable_fn per unique set of fetches (#27)
Creating a callable function in keras backend is the most expensive operation when we profile KerasHook.

https://github.com/tensorflow/tensorflow/blob/590d6eef7e91a6a7392c8ffffb7b58f2e0c8bc6b/tensorflow/python/keras/backend.py#L3473

if (self._callable_fn is None or feed_arrays != self._feed_arrays or
        symbol_vals != self._symbol_vals or
        feed_symbols != self._feed_symbols or self.fetches != self._fetches or
        session != self._session):
    # create callable
Whenever we ask for tensors to be saved, it forces Keras to recreate the callable, costing time. In this PR, we cache the callable_fn created by Keras for a given set of fetches, and the next time we try to add the same fetches, we update the callable_fn and change the previous fetches (exec_fn._fetches) to trick TF into not computing the callable.

We maintain this cache per mode, as different modes have different feeds (which is another of the conditions used to determine whether to compute the callable again)

Put this caching behavior under an environment variable so in case we run into any weird issues, we can turn off the cache.

This improves performance by 2.5x for ResNet50 training when saving weights, gradients, losses.