Compatibility with TF 2.7
...by fixing broken tests in test_keras_mirrored.py

Motive: Tensorflow 2.7.0 release

`smdebug.tensorflow.get_hook()` was returning `None`, which made many
tests fail, but this was actually a false alarm.
We just have to mark the version of tensorflow we are trying to release (`2.7.0`)
as a supported framework version, so `get_hook()` doesn't return
`None'...

I know what you think. It's not pythonic to silently fail and we should
instead throw an exception, but this code preceded me and I don't want
to introduce a breaking change (an exception being raised) and I don't
want to introduce warnings either in case `get_hook()' gets called
in unexpected ways multiple times during the lifecycle of a training job.

Instead, I just created a tests that alarms us of the situation
when `get_hook()` might return `None` when it's actually just because we forgot to update a constant.